Time: 1528 ms (5.00%) | Memory: 507.8 MB (5.01%) - LeetSync

+++ 18-4sum/4sum.cpp
@@ -17,7 +17,7 @@ public:
 					st.insert(tmp);
 				}
 
-				hashset.insert((long long)arr[k]);
+				hashset.insert(arr[k]);
 			}
 		}
 	}
